Fix memcached kill issue in each test case
1. There were 2 delete_all_buckets functions in BucketUtils - Fixed

2. We were calling the bucket delete task for all the existing buckets 
at the same time. 

I don't think cluster allows that. Moreover i don't see a use case for 
it.

3. Any such task doesn't exist anyways in any of the tasks files
(we have 2 tasks file as of now)

The call was present in the cluster.py and that is also removed as part 
of this change

This is all causing the BaseException in the basetestcase and hence 
memcached was getting killed in each test either passed or failed

Change-Id: I3c88a056f5b07d2643176677506728755bbae8bb
Reviewed-on: http://review.couchbase.org/109155
Reviewed-by: Ritesh Agarwal <ritesh.agarwal@couchbase.com>
Tested-by: Ritesh Agarwal <ritesh.agarwal@couchbase.com>