CBQE-6165: 1/n Add indexing + n1ql support to volume test
This adds support to create user defined number of secondary indexes on
collections after inital setup in the volume test. It creates equal
number of queries that are run during the entire test (during rebalance
/failover ops).

Also creating a new bucket spec for scalable stats. The goal for
scalable stats volume test is to generate high number of stats with the
help of large number of collections/scopes and indexes with less docs,
as number of docs don't significantly affect the number of stats
entities. So this bucket spec will have lesser docs but 1000
collections and 1000's of indexes.

Change-Id: I49b77db9cb0823ad3f5f6d3c462e39505dc12ced
Reviewed-on: http://review.couchbase.org/c/TAF/+/138626
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ashwin <ashwintrojan+github@gmail.com>