Check timestamps of memcached logs CRITICAL|ERROR|WARN messages
Currently, we fail a test even if we find the above messages in logs
from the previous tests. This is incorrect as we need to check if these
messages were from the current test.

To correct this we need to compare latest timesatamp from the grep's
results with the start of current test's timestamp. With this change,
if latest timestamp from log's grep results is less than the start of
test's timestamp we avoid failing the test.

Note: This would only work if slave's timest(timezone) matches that of
VMs' timezones

Change-Id: I9c6d2b710b00f20250b535961f9b8d2be96e4c24
Reviewed-on: http://review.couchbase.org/c/TAF/+/141229
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Balakumaran G <balakumaran.gopal@couchbase.com>
Reviewed-by: Balakumaran G <balakumaran.gopal@couchbase.com>