CBQE-6501: 2/n Rebalance task is not always the first dict
It is incorrect to assume that the first dict of the response from
pools/default/tasks contains the rebalance task. Orphan bucket
tasks have higher priority and hence they make take the first place
in the response dict.
Hence we iterate till we find the type=="rebalance" in the response
array

Change-Id: I735b3d587119eca10375bac9575e6e93f7917c9b
Reviewed-on: http://review.couchbase.org/c/TAF/+/148929
Reviewed-by: Balakumaran G <balakumaran.gopal@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>