CBQE-6995: 5/n Improve subdoc removes indices calculation
Consider this scenario, and what we are doing currently:
operation		doc-index		subdoc-index
insert 100 docs 	(0,100)			(0,0)
insert 100 subdocs	(0,100)			(0,100)
delete 20 subdocs	(0,100)			(20,100)

Now we cannot create 20 subdocs(even though 0 to 20 documents are
present) as end index of subdocs is already at 100.

To overcome this, we can delete (0,20) subdocs instead. In that way,
we will still have room for subsequent subdoc creates. So this
change, decrements the last index of subdocs object instead of
incrementing the first index of subdoc  object.

Change-Id: I825e49ce62167834d671ed0ad7fbd2ae69e85168
Reviewed-on: http://review.couchbase.org/c/TAF/+/156304
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ashwin <ashwintrojan+github@gmail.com>