CBQE-7092:remove get-coredumps check from testrunner.py in TAF.
It is already taken care in basetestcase teardown for each test.
We cannot do it after teardwon of the cluster as we need to make
certain decisions based on core found and critical messages in logs.

For example:
1. Data copy
2. Stop server on crash

If we do these things after teardown where buckets are deleted
and nodes are rebalance out we will loose the functionality.

Change-Id: Id22e07348582dfba3ea394688d011a611509460b
Reviewed-on: http://review.couchbase.org/c/TAF/+/157008
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Balakumaran G <balakumaran.gopal@couchbase.com>
Reviewed-by: Balakumaran G <balakumaran.gopal@couchbase.com>