CBQE-7129: [5/n] Add workloads
Add workloads for the creation and deletion of documents and documents
with xattributes. These workloads involve specifying the minimum and
maximum range of keys. This consequently allows the construction of
tests in which various workloads will change key ranges and test code
can then in turn reason about their effects.

Features functionality which allows user to introduce xattributes into
documents in specific vbuckets for rollback testing.

Change-Id: I145b36faf7deb11aad8b4b1181eebef86de2979b
Reviewed-on: http://review.couchbase.org/c/TAF/+/157614
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ashwin <ashwintrojan+github@gmail.com>