CBQE-7137: [8/n] Add deletion tests
Add testing for creating documents with xattributes and deleting them.
Expect the user attributes to be discarded and the system xattributes to
be accessible following deletion.

Adding testing for creating documents with xattributes, configuring
expiry by set the ttl for each document and waiting for the documents to
expire. Expect both the user and system xattributes to be inaccessible
following expiration.

Change-Id: I4ec0b9e6d97c30db452250f094abadbaab02f7a0
Reviewed-on: http://review.couchbase.org/c/TAF/+/157692
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ashwin <ashwintrojan+github@gmail.com>