CBQE-7141: [11/n] Rollback testing for xattributes
Add rollback testing for xattributes.

Add functionality for fetching actives on a specific node and for
validating the xattributes following rollback.

Context:
After stopping persistence, create documents with xattributes. Kill
memcached and wait for it to restart. At this point the replicas for
each active vbucket on 1 are further ahead. Similarly, the replica
vbuckets on node 1 are behind their active vbuckets.

Hence, the actives on the first node cause their replicas to rollback
and the replicas present on the first node to accept the new mutations
from their actives.

Change-Id: Iebfb2fa8f1455fa6881b2bd80023eacb181b0855
Reviewed-on: http://review.couchbase.org/c/TAF/+/157699
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ashwin <ashwintrojan+github@gmail.com>