CBQE-7188: Enable access_deleted create_as_deleted
Add the option of supplying the create_as_deleted and access_deleted
flags to the MutateInOptions so Tombstone documents can be directly
created and manipulated by sub-document operations.

Furthermore, enable access_deleted and create_as_deleted for each of the
sub-document operations accessible via async_load_gen_sub_docs including
insert, replace, upsert, remove. Enable access_deleted for the read
sub-document operation.

Add a missing store_semantics argument on the sub-document replace
operation which gives access to the CMD_SET operation which allows
supplying of a an path to change the entire value of a document.

Change-Id: Icfc830f5848ec4a86be69045e52ee4cf0311c35e
Reviewed-on: http://review.couchbase.org/c/TAF/+/160194
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ritesh Agarwal <ritesh.agarwal@couchbase.com>