CBQE-7278: Amend sub_doc_success test_basic_ops
Amended various errors in sub_doc_success's test_basic_ops.

Amended the success condition for the update operations to not rely on
hard-coded indices as they seem to have changed.

Amended an error in which the number of failures were added instead of
being subtracted to the count of updates causing the stat verification
to fail.

Similarly, amended an error in which the number of failures were
subtracted following deletion as opposed to being added.

Amended an error where the test was validating the remove operation to
have succeeded by checking it could read the attributes as opposed to
checking it could not read the attributes.

Made the assertion more accurate by accurately counting the expected
read failures for the delete operation.

Updated code to use self.assert prefixed methods as opposed to self.fail
to make the code easier to understand.

Amended an error where the failures in the insert operation were not
accounted for.

Amended an error where an empty dictionary was displayed as opposed to
the error message in the table of failed remove operation.

Amended an error where the test expects the sync writes committed
statistic to be populated on durability_level=None. However, this
statistic is only populated on durability_levels=majority, persist to
active and persist to majority.

Amended an error for update failures on read validation step in
sub_doc_success.test_basic_ops.

Added a doc-string and comments for clarity.

Change-Id: I44d3c4489048598b0aba7d5aeb8f4fbfc22ebc2d
Reviewed-on: http://review.couchbase.org/c/TAF/+/163085
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>
Reviewed-by: Ashwin <ashwintrojan+github@gmail.com>