CBQE-7334: ns_server's concurrent_requests
Added throughput for ns_server's concurrent requests. Used sockets to
open several connection's to one of ns_server's streaming api and
repeatedly read from each socket and attempt to keep them alive in a
round robin fashion.

The polling happen's in a seperate thread. The user can configure the
throughput which represents the number of concurrent connections for a
specific user to a specific node. This means we can produce throughput
for multiple users to various nodes in parallel.

The user can count the number of open connections which reflects the
number of successful concurrent requests and should be analogous to the
'cm_num_concurrent_requests' stat.

Change-Id: I2f22ece41d3473314500727644c87cb9f71c2633
Reviewed-on: https://review.couchbase.org/c/TAF/+/164622
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Balakumaran G <balakumaran.gopal@couchbase.com>