Add CouchbaseException to catch block
Add CouchbaseException to the list of caught exceptions in the upsert
operation. This exception is thrown by the SDK when the status code is
unknown, for example in the case of limits exceeded. Interestingly, this
is not caught by the final 'except Exception' block as I imagine
CouchbaseException from the world of Java is not a child of Python's
Exception.

Change-Id: I5d964f61fc0e5c8626e0daced3709818f41091f9
Reviewed-on: https://review.couchbase.org/c/TAF/+/166909
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ashwin <ashwin.trojan@gmail.com>