Re-enable error checking and reduce log file size
Added a retry-with-timeout around error validation, this allows
validation to succeed at the moment the limit is exceeded. Previously
the validation failed as the exceeded limit gets reset periodically
every 60 seconds meaning the error message was not raised.

Added a flag `log_errors` to RestConnection. The throughput producers
create a large number of http requests. When the ns_server limits are
exceeded, the error logs produced are very large. Setting this flag
prevents this from happening.

Reduced the number of nodes to improve stability of ns_server_ingress
tests.

Change-Id: I65ee7a54e3e87414ef7bb3b538a9e3ec8c942f6f
Reviewed-on: https://review.couchbase.org/c/TAF/+/168120
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <ankush.sharma@couchbase.com>