CBQE-7879: Dispatcher: Rerunning failed job is rerunning the entire suite instead of just failed ones
Fixed the rerun scripts to rerun in case of
cloud runs too. These runs have parameters
without any values, which is causing the
reruns to fail.
Also added logic to add bucket storage and
GSI types to the tests to differentiate same
jobs but with differing bucket storage or
gsi types.

Change-Id: I455814d53935184b2b0f2ba8c3f3c35c05656ad8
Reviewed-on: https://review.couchbase.org/c/TAF/+/190020
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ashwin <ashwin.govindarajulu@couchbase.com>