CBQE-8104 : Function wait_for_rebalance_task_to_complete resets cluster.servers based on a bool value
The rebalance task requires the cluster.servers to be populated fully. wait_for_rebalance_task_to_complete now has a param called reset_cluster_servers. The cluster.servers is only reset in the function if the value of reset_cluster_servers is True. Test cases which require the cluster.servers to be fully populated and not reset can pass the param and have the cluster.servers unchanged. Consequently the cbas upgrades now call the function wait_for_rebalance_task_to_complete with reset_cluster_servers=False such that the cluster.servers is not reset

Change-Id: I2e72fb6f69b19a4a2a17acca2830089a73fb19e6
Reviewed-on: https://review.couchbase.org/c/TAF/+/204439
Tested-by: Raghav S K <raghav.sk@couchbase.com>
Reviewed-by: Ashwin <ashwin.govindarajulu@couchbase.com>
Reviewed-by: Umang Agrawal <umang.agrawal@couchbase.com>
Reviewed-by: Raghav S K <raghav.sk@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>