Adding a cleanup of all nodes in cbas_upgrade teardown
Teardown is called before cleanup as we need to collect cbcollect logs. If cleanup is called before teardown, cbcollect will capture only one node logs, with no buckets and no rebalance hangs. During rebalance tests in upgrades, the first node (which usually has kv service) can be added back as cbas. This causes the setup in the succeeding test case to fail as it cannot rebalance out all nodes. The way to handle that is to cleanup in this test case and then proceed to the next

Change-Id: I812471ebb9b47d06685d14a97cb6162bfc3ea93b
Reviewed-on: https://review.couchbase.org/c/TAF/+/205993
Reviewed-by: Balakumaran G <balakumaran.gopal@couchbase.com>
Tested-by: Raghav S K <raghav.sk@couchbase.com>
Reviewed-by: Raghav S K <raghav.sk@couchbase.com>