Fixed the download of previous test results
Fixed the download of previous test results while
merging the xmls while running the rerun script.
The issue was that the download was returning as
as string and string cannot be decoded in latest
python versions. Added fix for this. Added extra
checks incase the download returns as bytes instead.

Change-Id: Ia6843d974bcf47b5d26ac99ed64f258831b0cb67
Reviewed-on: https://review.couchbase.org/c/TAF/+/213214
Reviewed-by: Ashwin <ashwin.govindarajulu@couchbase.com>
Tested-by: Ritam Sharma <ritam@couchbase.com>
Reviewed-by: Ritam Sharma <ritam@couchbase.com>