Adding doc_ops=create in doc loader call
In some Magma cases, init_loading in magma_base is skipped and instead executed in the test function.

Doc Loader takes doc_ops as one of its params to divide the doc_op_percent based on it. Problem arised when we explicitly pass doc_ops in the test case (actually meant for subsequent load), which interfered with the first creates. To avoid this, we don't rely on self.doc_ops for the first creates, instead we pass doc_ops="create".

Change-Id: Ia43d2010028b869870438ee5c44aaaa72916f5df
Reviewed-on: https://review.couchbase.org/c/TAF/+/222812
Reviewed-by: <ankush.sharma@couchbase.com>
Reviewed-by: Ashwin <ashwin.govindarajulu@couchbase.com>
Tested-by: Vibhav S P <vibhav.sp@couchbase.com>